+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.0.1
+- Added debug commands
+- Added `owner_id` field to settings
+
 ### 1.0.0
 - Added BanUtils
 - Switched database to tortoise-orm